MAY 29, 2016 CANSOLO PACKET TRANSMISSION TEST
Updated the Honeywell pressure sensor driver to include some extra members
for data storage. Now there is a private member to store the latest
pressure reading for easier access, this may not be entirely necessary
with the current configuration but may become useful in the future.
Expanded upon the init function to perform useful communication protocol
assignments on the i2c bus. implemented a function that uses the dynamic
pressure equation to return the current speed given the latest pressure
reading. For most accurate results call this function immediately after
performing a read.
[WIP]

TODO:
	Test new init function on hardware,
	Test velocity conversion on hardware using compressed air can

 On branch packet
	modified:   lib/ABP/ABP.cpp
	modified:   lib/ABP/ABP.hpp